Remove legacy code for handling downstream components not being ready
Checking for downstream readiness was initially a problem as described in #40. That was due to initial packets being sent before the network is ready. It's partially solved in #69 when initial packets are only sent upon the respective network has been loaded. This PR solves the other half which is initial packets being sent in order of the programmer's specification so by the time the network is run, everything should be ready anyway.