Add all detected components but not those already added
NoFlo currently ignores entire packages of the same name that have already been added. This creates a problem as the graph's number of dependencies increases as there's bound to be some packages using newer versions of the same package than others. Right now it imports the first one as returned by `read-installed` so an older package may shadow a newer version of itself. This patch avoids conflicts at the component-level rather than the package-level.